Breaking apart class change source into compiler level vs registry level
Breaking out class from file changes
