+++ module/compiler/index.ts
@@ -1,6 +1,7 @@
 export * from './src/util';
 export * from './src/source';
 export * from './src/compiler';
+export * from './src/file-source';
 export * from './src/transformer/manager';
 export * from './src/transformer/types';
 export * from './src/transformer/state';
